requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
pymongo==3.12.0
dnspython
python-dotenv